+++ client/app/components/composites/AvatarDropdown/AvatarDropdown.css
@@ -1,10 +1,6 @@
-.avatar {
+.avatarDropdown {
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: flex-start;
-}
-
-.avatarImage {
-  border-radius: 50%;
 }
