+++ client/app/components/elements/Avatar/Avatar.css
@@ -1,0 +1,3 @@
+.avatar {
+  border-radius: 50%;
+}
